# perf: loops optimization