--- a/lib/client.js
+++ b/lib/client.js
@@ -358,7 +358,9 @@ class Client extends Dispatcher {
       err = new ClientDestroyedError()
     }
 
-    for (const request of this[kQueue].splice(this[kPendingIdx])) {
+    const requests = this[kQueue].splice(this[kPendingIdx])
+    for (let i = 0; i < requests.length; i++) {
+      const request = requests[i]
       request.onError(err)
       assert(request.aborted)
     }
@@ -370,8 +372,8 @@ class Client extends Dispatcher {
     const onDestroyed = () => {
       const callbacks = this[kOnDestroyed]
       this[kOnDestroyed] = null
-      for (const callback of callbacks) {
-        callback(null, null)
+      for (let i = 0; i < callbacks.length; i++) {
+        callbacks[i](null, null)
       }
     }
 
@@ -706,14 +708,17 @@ class Parser {
     let keepAlive
     let trailers
 
-    for (let n = 0; n < rawHeaders.length; n += 2) {
-      const key = rawHeaders[n + 0]
+    let looking = true
+    for (let n = 0; n < rawHeaders.length && looking; n += 2) {
+      const key = rawHeaders[n]
       const val = rawHeaders[n + 1]
 
       if (!keepAlive && key.length === 10 && key.toString().toLowerCase() === 'keep-alive') {
         keepAlive = val
+        looking = !trailers
       } else if (!trailers && key.length === 7 && key.toString().toLowerCase() === 'trailer') {
         trailers = val
+        looking = !keepAlive
       }
     }
 
@@ -814,10 +819,11 @@ class Parser {
       return
     }
 
-    for (const trailer of trailers) {
+    for (let i = 0; i < trailers.length; i++) {
+      const trailer = trailers[i]
       let found = false
       for (let n = 0; n < rawTrailers.length; n += 2) {
-        const key = rawTrailers[n + 0]
+        const key = rawTrailers[n]
         if (key.length === trailer.length && key.toString().toLowerCase() === trailer.toLowerCase()) {
           found = true
           break
@@ -912,7 +918,10 @@ function onSocketError (err) {
     assert(client[kPendingIdx] === client[kRunningIdx])
     // Error is not caused by running request and not a recoverable
     // socket error.
-    for (const request of client[kQueue].splice(client[kRunningIdx])) {
+
+    const requests = client[kQueue].splice(client[kRunningIdx])
+    for (let i = 0; i < requests.length; i++) {
+      const request = requests[i]
       request.onError(err)
       assert(request.aborted)
     }
@@ -955,7 +964,9 @@ function onSocketClose () {
     assert(client[kPending] === 0)
 
     // Fail entire queue.
-    for (const request of client[kQueue].splice(client[kRunningIdx])) {
+    const requests = client[kQueue].splice(client[kRunningIdx])
+    for (let i = 0; i < requests.length; i++) {
+      const request = requests[i]
       request.onError(err)
       assert(request.aborted)
     }