--- a/lib/core/request.js
+++ b/lib/core/request.js
@@ -83,11 +83,13 @@ class Request {
         throw new InvalidArgumentError('headers array must be even')
       }
       for (let i = 0; i < headers.length; i += 2) {
-        processHeader(this, headers[i + 0], headers[i + 1])
+        processHeader(this, headers[i], headers[i + 1])
       }
     } else if (headers && typeof headers === 'object') {
-      for (const [key, val] of Object.entries(headers)) {
-        processHeader(this, key, val)
+      const keys = Object.keys(headers)
+      for (let i = 0; i < keys.length; i++) {
+        const key = keys[i]
+        processHeader(this, key, headers[key])
       }
     } else if (headers != null) {
       throw new InvalidArgumentError('headers must be an object or an array')