--- a/lib/handler/redirect.js
+++ b/lib/handler/redirect.js
@@ -125,9 +125,11 @@ function cleanRequestHeaders (headers, removeContent) {
       }
     }
   } else if (headers && typeof headers === 'object') {
-    for (const [key, val] of Object.entries(headers)) {
+    const keys = Object.keys(headers)
+    for (let i = 0; i < keys.length; i++) {
+      const key = keys[i]
       if (!shouldRemoveHeader(key, removeContent)) {
-        ret.push(key, val)
+        ret.push(key, headers[key])
       }
     }
   } else if (headers != null) {