npm run bench this branch

[1] http - no agent  x 711 ops/sec ±0.86% (84 runs sampled)
[1] http - keepalive x 760 ops/sec ±0.64% (78 runs sampled)
[1] http - keepalive - multiple sockets x 4,732 ops/sec ±1.70% (84 runs sampled)
[1] undici - pipeline x 5,531 ops/sec ±1.81% (84 runs sampled)
[1] undici - request x 5,656 ops/sec ±0.77% (83 runs sampled)
[1] undici - pool - request - multiple sockets x 5,720 ops/sec ±0.64% (86 runs sampled)
[1] undici - stream x 5,809 ops/sec ±0.60% (84 runs sampled)
[1] undici - dispatch x 5,986 ops/sec ±0.45% (88 runs sampled)
[1] undici - noop x 7,414 ops/sec ±0.34% (87 runs sampled)

npm run bench main branch

[1] http - no agent  x 719 ops/sec ±0.73% (84 runs sampled)
[1] http - keepalive x 747 ops/sec ±0.52% (79 runs sampled)
[1] http - keepalive - multiple sockets x 4,644 ops/sec ±1.91% (81 runs sampled)
[1] undici - pipeline x 5,437 ops/sec ±2.06% (80 runs sampled)
[1] undici - request x 5,624 ops/sec ±1.01% (84 runs sampled)
[1] undici - pool - request - multiple sockets x 5,639 ops/sec ±0.59% (85 runs sampled)
[1] undici - stream x 5,856 ops/sec ±0.72% (85 runs sampled)
[1] undici - dispatch x 6,046 ops/sec ±0.58% (86 runs sampled)
[1] undici - noop x 7,342 ops/sec ±0.28% (89 runs sampled)