@simone-sanfratello 2021-04-02
> Can you check if this is giving a warning from `standard`?

yes, it is. Should I use eslint and extend standard, then allow vars?