@ronag 2021-04-02
Is var actually faster? Do we have any numbers?