@simone-sanfratello 2021-04-02
> Can you check if this is giving a warning from `standard`?

with standard `v.14`, used in the project, is fine, but with the `v.16`, the latest, is not