@szmarczak 2021-04-09
> Replace all i++ 's with ++i as it is faster and more memory efficient (doesn't make a copy)

@tbnritzdoge I'd have never thought of that 🤯 Nice spot.