@ronag 2021-04-09
> Replace all i++ 's with ++i as it is faster and more memory efficient (doesn't make a copy)

That’s somewhat true with c++ iterators but doesn’t really apple here.