@szmarczak 2021-04-09
> That’s somewhat true with c++ iterators but doesn’t really apple here.

I guess V8 optimizes this?