@Nytelife26 2021-04-09
```
prefix x 522,745,219 ops/sec ±1.32% (85 runs sampled)
suffix x 516,346,780 ops/sec ±1.19% (84 runs sampled)
```
Having ran a benchmark between the two, `i++` is slightly slower usually. 

However, having ran another benchmark,

> please use `for (var`, it's faster.

```
for-of x 68,941,007 ops/sec ±1.07% (87 runs sampled)
c-style (prefix) x 92,530,979 ops/sec ±0.57% (96 runs sampled)
c-style (suffix) x 92,918,425 ops/sec ±0.45% (97 runs sampled)
```

for-of turns out to be significantly slower. Aside from that, it is largely irrelevant which is used, be it prefix or suffix, because the performance difference is inconsistent and negligible. It may be prevalent to use the prefix method though, given that, as Doge rightly pointed out, it creates an extra copy on the stack.