@simone-sanfratello 2021-04-09
> for-of turns out to be significantly slower. Aside from that, it is largely irrelevant which is used, be it prefix or suffix, because the performance difference is inconsistent and negligible. It may be prevalent to use the prefix method though, given that, as Doge rightly pointed out, suffix creates an extra copy on the stack which leads to unnecessary memory usage.

I'm doing some (a lot) benchmarks too, yes for-of is quite slower than "classic" for
I'll add the `++i` to the cases, but looks like it doesn't make much difference