@wzhouwzhou 2021-04-10
Agreed with @ronag, as usage of ++i and i++ here is likely optimised away with return value dropped, and any turbofan v8 loop [unrolling](https://github.com/v8/v8/blob/dc712da548c7fb433caed56af9a021d964952728/src/codegen/code-stub-assembler.h#L3706) will also condense that down into at most a couple of instructions. However, those optimisations may only occur in some picture-perfect cases. 

At this point, we should always consider whether we are venturing into premature optimisation territory, as with web-oriented applications and libraries, the bottlenecks will almost certainly occur elsewhere. I would hate for a library to become any less maintainable or inconsistent with itself (i.e., i++'s here, ++i's there, lack of structure and order) over changes like that. (Same issue as the fastify reference above).