@tbnritzdoge 2021-04-10
- [ ] You should note that i++ and ++i are  actually not the same thing and are different things hence force the different optimization levels. In addition to your point I think anyone with a interest in these libraries has a obvious wish to reach the pinnacle of performance and these incremental changes push us to that edge, but if your so insistent on your techniques you should consider using slower and "cleaner" alternatives such as the built in http(s) and express libraries @wzhouwzhou