@wzhouwzhou 2021-04-10
@tbnritzdoge We're really gratefulfor your interest in making this library faster! As mentioned previously, we see contradictory test results with no obvious improvements. That being said, for those wishing to reach the actual "pinnacle of performance" I recommend you learn languages such as C++ and rust! One of the most important but overlooked aspects of library maintenance and open source design involves maintainability, a skill that takes time and real-world exposure for you to experience. 
Looking forward to see your growth in the future!