@tbnritzdoge 2021-04-10
> @tbnritzdoge We're really grateful for your interest in making this library faster! As mentioned previously, we see contradictory test results with no obvious improvements. That being said, for those wishing to reach the actual "pinnacle of performance" I recommend you learn languages such as C++ and rust! One of the most important but overlooked aspects of library maintenance and open source design involves maintainability, a skill that takes time and real-world exposure for you to experience. 
> 
> Looking forward to see your growth in the future!

1. I don't see any memory benchmarks nor do I see a understanding of node's internals as its very obvious that having a non copying function is better than a copying one.

2. I do indeed spend a lot of my time writing rust, but if you worked for a real company you would know that porting an entire codebase is a immense chore so it's better in the short term to find simple optimizations on what you have rather than scrap everything and rewrite to a new language.