@tbnritzdoge 2021-04-10
> 1. You should take a look into how ignition and turbofan really work - I recommend you start with some v8 blogs available on google. Given any impressive rust background, the conversion from bytecode to machine code should be easy for you.
> 
> 2. The overhead of porting codebases is irrelevant to the conversation - if you worked at a real company you would know that we are talking about a series of tradeoffs and performance on orders of magnitudes worth of difference, and that the juxtaposition is the critical focus, rather than the execution. Both enterprise level code and typical users of NodeJS use scripting language for ease of use. It's better in the short-term for you to focus on areas of the library that actually need improvement to design and control flow with known bottlenecks rather than make a big deal out of a subjective convention with insufficient proof.

If you actually were as proficient as you act you are you would know the performance benefits.