@wzhouwzhou 2021-04-10
At this point it seems that you've entirely strayed away from any argument grounded in theory and reason - understanding proficiency as well as the multiple facets of performance myself (while taking into account developer efficiency), it looks like there is nothing more to be discussed with you in your current state. Please go touch some grass and consider what gives you that right.

For what it's worth, the decision on this PR has already been made already, so it's confusing why you belabour your suggestions and express significant amounts of discontent at any discussions regarding them. It's childish and unprofessional. We are just having a light chat about the pros and cons.

Kindly don't make an immense chore out of this and any future contributions to NodeJS in the future. Thanks!